# api.yml
...
openapi: "3.0.0"
info:
  version: 1.0.0
  title: User Service
  license:
    name: MIT
servers:
  - url: http://localhost
paths:
  /estate:
    post:
      summary: Create a new estate
      operationId: CreateEstate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EstateRequest"
      responses:
        "201":
          description: Estate created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EstateResponse"
        "400":
          description: Invalid request payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
    ErrorResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
    EstateRequest:
      type: object
      required:
        - length
        - width
      properties:
        length:
          type: integer
          example: 10
        width:
          type: integer
          example: 10
    EstateResponse:
      type: object
      required:
        - id
      properties:
        id:
          type: string
          format: string
          example: "550e8400-e29b-41d4-a716-446655440000"
...